Extract sign-in form to login page
