+++ src/login-page.jsx
@@ -1,0 +1,63 @@
+import React, { useState } from "react";
+import {
+	signInWithPopup,
+	signInWithEmailAndPassword,
+} from "firebase/auth";
+import { Link, useNavigate } from "react-router-dom";
+import { Button } from "./components/ui/ui";
+import { auth, provider } from "./firebase";
+
+export default function LoginPage() {
+	const [error, setError] = useState("");
+	const [email, setEmail] = useState("");
+	const [password, setPassword] = useState("");
+	const navigate = useNavigate();
+
+	const handleGoogleSignIn = async () => {
+		try {
+			setError("");
+			await signInWithPopup(auth, provider);
+			navigate("/");
+		} catch (e) {
+			console.error("sign-in failed", e);
+			setError("Sign in failed");
+		}
+	};
+
+	const handleEmailSignIn = async () => {
+		try {
+			setError("");
+			await signInWithEmailAndPassword(auth, email, password);
+			navigate("/");
+		} catch (e) {
+			console.error("email sign-in failed", e);
+			setError("Sign in failed");
+		}
+	};
+
+	return (
+		<div className="auth-panel login-page">
+			<h2>Login</h2>
+			<input
+				className="auth-input"
+				type="email"
+				placeholder="Email"
+				value={email}
+				onChange={(e) => setEmail(e.target.value)}
+			/>
+			<input
+				className="auth-input"
+				type="password"
+				placeholder="Password"
+				value={password}
+				onChange={(e) => setPassword(e.target.value)}
+			/>
+			<Button onClick={handleEmailSignIn}>Login</Button>
+			<Button onClick={handleGoogleSignIn}>Google</Button>
+			<Link to="/profile/new" className="auth-link">
+				Create profile
+			</Link>
+			{error && <span className="auth-error">{error}</span>}
+		</div>
+	);
+}
